+++ packages/ai/src/agent/agent.ts
@@ -15,7 +15,6 @@ import { CallSettings } from '../prompt/call-settings';
 import { Prompt } from '../prompt/prompt';
 import { TelemetrySettings } from '../telemetry/telemetry-settings';
 import { LanguageModel, ToolChoice } from '../types/language-model';
-import { ProviderMetadata } from '../types/provider-metadata';
 import { convertToModelMessages } from '../ui/convert-to-model-messages';
 import { InferUITools, UIMessage } from '../ui/ui-messages';
 
@@ -24,6 +23,11 @@ export type AgentSettings<
   OUTPUT = never,
   OUTPUT_PARTIAL = never,
 > = CallSettings & {
+  /**
+   * The name of the agent.
+   */
+  name?: string;
+
   /**
    * The system prompt to use.
    */
@@ -91,6 +95,13 @@ A function that attempts to repair a tool call that failed to parse.
   onStepFinish?: GenerateTextOnStepFinishCallback<NoInfer<TOOLS>>;
 
   /**
+Additional provider-specific options. They are passed through
+to the provider from the AI SDK and enable provider-specific
+functionality that can be fully encapsulated in the provider.
+         */
+  providerOptions?: ProviderOptions;
+
+  /**
    * Context that is passed into tool calls.
    *
    * Experimental (can break in patch releases).
@@ -119,45 +130,25 @@ export class Agent<
     this.settings = settings;
   }
 
+  /**
+   * The name of the agent.
+   */
+  get name(): string | undefined {
+    return this.settings.name;
+  }
+
+  /**
+   * The tools that the agent can use.
+   */
   get tools(): TOOLS {
     return this.settings.tools as TOOLS;
   }
 
-  async generate(
-    options: Prompt & {
-      /**
-Additional provider-specific metadata. They are passed through
-from the provider to the AI SDK and enable provider-specific
-results that can be fully encapsulated in the provider.
-   */
-      providerMetadata?: ProviderMetadata;
-      /**
-Additional provider-specific metadata. They are passed through
-to the provider from the AI SDK and enable provider-specific
-functionality that can be fully encapsulated in the provider.
-         */
-      providerOptions?: ProviderOptions;
-    },
-  ): Promise<GenerateTextResult<TOOLS, OUTPUT>> {
+  async generate(options: Prompt): Promise<GenerateTextResult<TOOLS, OUTPUT>> {
     return generateText({ ...this.settings, ...options });
   }
 
-  stream(
-    options: Prompt & {
-      /**
-Additional provider-specific metadata. They are passed through
-from the provider to the AI SDK and enable provider-specific
-results that can be fully encapsulated in the provider.
-   */
-      providerMetadata?: ProviderMetadata;
-      /**
-Additional provider-specific metadata. They are passed through
-to the provider from the AI SDK and enable provider-specific
-functionality that can be fully encapsulated in the provider.
-         */
-      providerOptions?: ProviderOptions;
-    },
-  ): StreamTextResult<TOOLS, OUTPUT_PARTIAL> {
+  stream(options: Prompt): StreamTextResult<TOOLS, OUTPUT_PARTIAL> {
     return streamText({ ...this.settings, ...options });
   }