+++ packages/openai/src/responses/openai-responses-api-types.ts
@@ -9,6 +9,8 @@ export type OpenAIResponsesInputItem =
   | OpenAIResponsesFunctionCall
   | OpenAIResponsesFunctionCallOutput
   | OpenAIResponsesComputerCall
+  | OpenAIResponsesLocalShellCall
+  | OpenAIResponsesLocalShellCallOutput
   | OpenAIResponsesReasoning
   | OpenAIResponsesItemReference;
 
@@ -67,6 +69,26 @@ export type OpenAIResponsesComputerCall = {
   type: 'computer_call';
   id: string;
   status?: string;
+};
+
+export type OpenAIResponsesLocalShellCall = {
+  type: 'local_shell_call';
+  id: string;
+  call_id: string;
+  action: {
+    type: 'exec';
+    command: string[];
+    timeout_ms?: number;
+    user?: string;
+    working_directory?: string;
+    env?: Record<string, string>;
+  };
+};
+
+export type OpenAIResponsesLocalShellCallOutput = {
+  type: 'local_shell_call_output';
+  call_id: string;
+  output: string;
 };
 
 export type OpenAIResponsesItemReference = {
@@ -177,8 +199,12 @@ export type OpenAIResponsesTool =
       moderation: 'auto' | undefined;
       output_compression: number | undefined;
       output_format: 'png' | 'jpeg' | 'webp' | undefined;
+      partial_images: number | undefined;
       quality: 'auto' | 'low' | 'medium' | 'high' | undefined;
       size: 'auto' | '1024x1024' | '1024x1536' | '1536x1024' | undefined;
+    }
+  | {
+      type: 'local_shell';
     };
 
 export type OpenAIResponsesReasoning = {