+++ packages/rsc/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,33 @@
 
 ### Patch Changes
 
+- Updated dependencies [0adc679]
+- Updated dependencies [9b8d17e]
+  - ai@5.1.0-beta.18
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- ai@5.1.0-beta.17
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [14ca35d]
+  - ai@5.1.0-beta.16
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- ai@5.1.0-beta.15
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
 - Updated dependencies [7c3c216]
 - Updated dependencies [8dac895]
 - Updated dependencies [e0d1ea9]